back/python/image_storage/requirements.txt
fastapi~=0.110.2
uvicorn~=0.29.0
python-dotenv~=1.0.1
asyncio~=3.4.3
pymongo~=4.6.3
pydantic~=2.6.4
httpx~=0.27.0
python-multipart~=0.0.9
 pillow~=10.3.0